% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chr_distribution_plot}
\alias{chr_distribution_plot}
\title{Chromosome distribution plot}
\usage{
chr_distribution_plot(data, outliers = T, show_sign = F)
}
\arguments{
\item{data}{Output of one of the integration functions or output of
\strong{run_multiomics}}

\item{outliers}{logical. Should outliers be showed in the plot. Default is
set to TRUE}

\item{show_sign}{logical. Should barplot colors be defined according to the
sign of the coefficients ? In this case only significant coefficients will
be included in the plot. Default set to FALSE.}
}
\description{
The chromosome distribution plot is a barplot showing the number of
significant/non significant or positive/negative coefficients
for each chromosome.
}
