% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{circos_plot}
\alias{circos_plot}
\title{Circos plot}
\usage{
circos_plot(model_results, species = "hg38")
}
\arguments{
\item{model_results}{Output of an integration model. Not available for
the entire \strong{run_multiomics} output. You can provide one of the results
of the MultiOmics object (e.g. muliomics$gene_cnv_res) or provide the
results of one of the other integration functions.}

\item{species}{Species information, default is "hg38". See
\strong{circos.initializeWithIdeogram} function from \strong{circlize} package for
further information}
}
\description{
This function will generate a Circos plot for an integration model. It will
display in the first two layer the data of the covariates and of the
response variable and when possible in the third layer the values of the
significant coefficients
}
