% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_models.R
\name{.allgene_edgeR_model}
\alias{.allgene_edgeR_model}
\title{All edgeR gene models}
\usage{
.allgene_edgeR_model(
  response_var,
  design_mat_allgene = NULL,
  offset_allgene = NULL,
  norm_method = "TMM"
)
}
\description{
All edgeR gene models
}
