% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.build_histo_TFbyChr}
\alias{.build_histo_TFbyChr}
\title{Build histogram for TFs/miRNAs by chromosome}
\usage{
.build_histo_TFbyChr(histo_data)
}
\description{
Build histogram for TFs/miRNAs by chromosome
}
