% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.building_result_matrices}
\alias{.building_result_matrices}
\title{Building result matrices}
\usage{
.building_result_matrices(model_results, type, single_cov = FALSE)
}
\description{
Building result matrices
}
