% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.check_reactive_bg_enrich}
\alias{.check_reactive_bg_enrich}
\title{Check Reactive Background Enrichment}
\usage{
.check_reactive_bg_enrich(bg_enrich, input, output, session)
}
\description{
Check Reactive Background Enrichment
}
