% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.circos_track_cols}
\alias{.circos_track_cols}
\title{Generate color palette for Circos track}
\usage{
.circos_track_cols(vvalues, n = 50, colors = NULL)
}
\description{
Generate color palette for Circos track
}
