% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.covariates_check}
\alias{.covariates_check}
\title{Covariates check}
\usage{
.covariates_check(
  response_var,
  covariates,
  interactions,
  steady_covariates = NULL,
  linear = FALSE,
  reference = NULL
)
}
\description{
Covariates check
}
