% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.create_coef_track}
\alias{.create_coef_track}
\title{Create Coefficient track for Circos visualization}
\usage{
.create_coef_track(gr)
}
\description{
Create Coefficient track for Circos visualization
}
