% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.create_composed_view}
\alias{.create_composed_view}
\title{Create composed view for Circos visualization}
\usage{
.create_composed_view(tracks, width, height)
}
\description{
Create composed view for Circos visualization
}
