% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.create_expr_track}
\alias{.create_expr_track}
\title{Create Expression track for Circos visualization}
\usage{
.create_expr_track(gr, genomic = F)
}
\description{
Create Expression track for Circos visualization
}
