% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.create_single_track}
\alias{.create_single_track}
\title{Create single track for Circos visualization}
\usage{
.create_single_track(
  data,
  dataValue = NULL,
  x_axis = NULL,
  xe_axis = NULL,
  y_axis = NULL,
  colorField = NULL,
  colorDomain = NULL,
  colorRange = NULL,
  tooltipField1 = NULL,
  tooltipTitle = NULL,
  tooltipAlt1 = NULL,
  tooltipField2 = NULL,
  tooltipAlt2 = NULL,
  tooltipField3 = NULL,
  tooltipAlt3 = NULL,
  tooltipField4 = NULL,
  tooltipAlt4 = NULL,
  tooltipField5 = NULL,
  tooltipAlt5 = NULL,
  legend = NULL,
  colorType = NULL,
  title = NULL
)
}
\description{
Create single track for Circos visualization
}
