% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{.def_cnv_integration}
\alias{.def_cnv_integration}
\title{cnv integration}
\usage{
.def_cnv_integration(
  expression,
  cnv_data,
  sequencing_data,
  normalize,
  norm_method,
  BPPARAM,
  ...
)
}
\description{
cnv integration
}
