% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{.def_enrich}
\alias{.def_enrich}
\title{Enrichment}
\usage{
.def_enrich(
  data,
  species,
  pvalueCutoff,
  pAdjustMethod,
  qvalueCutoff,
  ont,
  run_go = TRUE,
  run_kegg = TRUE,
  run_reactome = FALSE,
  ...
)
}
\description{
Enrichment
}
