% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{.def_genomic_integration}
\alias{.def_genomic_integration}
\title{def genomic integration}
\usage{
.def_genomic_integration(
  expression,
  cnv_data,
  methylation,
  sequencing_data,
  normalize,
  norm_method,
  interactions,
  scale,
  BPPARAM,
  ...
)
}
\description{
def genomic integration
}
