% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{.def_met_integration}
\alias{.def_met_integration}
\title{defining met integration}
\usage{
.def_met_integration(
  expression,
  methylation,
  sequencing_data,
  normalize,
  norm_method,
  BPPARAM,
  ...
)
}
\description{
defining met integration
}
