% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{.def_tf_integration}
\alias{.def_tf_integration}
\title{def tf integration}
\usage{
.def_tf_integration(
  expression,
  tf_expression,
  interactions,
  type,
  sequencing_data,
  species,
  normalize,
  norm_method,
  normalize_cov,
  norm_method_cov,
  BPPARAM,
  ...
)
}
\description{
def tf integration
}
