% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.download_csv}
\alias{.download_csv}
\title{Handle CSV file download}
\usage{
.download_csv(
  tf = FALSE,
  deg = FALSE,
  bg_enr = NULL,
  plotType = "histo",
  input,
  output,
  data_table,
  i = NULL
)
}
\description{
Handle CSV file download
}
