% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{.download_gene_info_biomart}
\alias{.download_gene_info_biomart}
\title{Download gene infos from biomaRt}
\usage{
.download_gene_info_biomart(
  genes,
  species = "hsa",
  filters = c("hgnc_symbol", "ensembl_gene_id", "entrezgene_id"),
  ...
)
}
\description{
Download gene infos from biomaRt
}
