% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_chrDistrib_box}
\alias{.gint_chrDistrib_box}
\title{Defining Chromosome distribution box description}
\usage{
.gint_chrDistrib_box()
}
\description{
Defining Chromosome distribution box description
}
