% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_coefDistrib_box}
\alias{.gint_coefDistrib_box}
\title{Defining coefficients distribution box description}
\usage{
.gint_coefDistrib_box()
}
\description{
Defining coefficients distribution box description
}
