% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_chrDistribDEGs}
\alias{.gint_subItem_chrDistribDEGs}
\title{Chromosome Distribution Subitem for class comparison}
\usage{
.gint_subItem_chrDistribDEGs(data_table)
}
\description{
Chromosome Distribution Subitem for class comparison
}
