% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_chrDistribGenomic}
\alias{.gint_subItem_chrDistribGenomic}
\title{Chr Distribution Subitem for Genomic Integration}
\usage{
.gint_subItem_chrDistribGenomic(data_table)
}
\description{
Chr Distribution Subitem for Genomic Integration
}
