% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_chrDistribTranscript}
\alias{.gint_subItem_chrDistribTranscript}
\title{Chr Distribution Subitem for Transcription Integration}
\usage{
.gint_subItem_chrDistribTranscript(data_table)
}
\description{
Chr Distribution Subitem for Transcription Integration
}
