% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_circosCompleteInt}
\alias{.gint_subItem_circosCompleteInt}
\title{Circos Subitem for Complete Integration}
\usage{
.gint_subItem_circosCompleteInt(data_table)
}
\description{
Circos Subitem for Complete Integration
}
