% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_coefDistribDEGs}
\alias{.gint_subItem_coefDistribDEGs}
\title{Coefficients Distribution Subitem for Class Comparison}
\usage{
.gint_subItem_coefDistribDEGs(data_table)
}
\description{
Coefficients Distribution Subitem for Class Comparison
}
