% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_coefDistribGenomic}
\alias{.gint_subItem_coefDistribGenomic}
\title{Coefficients Distribution Subitem for Genomic Integration}
\usage{
.gint_subItem_coefDistribGenomic(data_table)
}
\description{
Coefficients Distribution Subitem for Genomic Integration
}
