% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_coefDistribTranscript}
\alias{.gint_subItem_coefDistribTranscript}
\title{Coefficients Distribution Subitem for Transcription Integration}
\usage{
.gint_subItem_coefDistribTranscript(data_table)
}
\description{
Coefficients Distribution Subitem for Transcription Integration
}
