% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_enrichTranscript}
\alias{.gint_subItem_enrichTranscript}
\title{Enrichment Subitem for Transcription Integration}
\usage{
.gint_subItem_enrichTranscript(data_table)
}
\description{
Enrichment Subitem for Transcription Integration
}
