% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_HeatmapDEGs}
\alias{.gint_subItem_HeatmapDEGs}
\title{Heatmap Subitem for class comparison}
\usage{
.gint_subItem_HeatmapDEGs(data_table)
}
\description{
Heatmap Subitem for class comparison
}
