% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_HeatmapGenomic}
\alias{.gint_subItem_HeatmapGenomic}
\title{Heatmap Subitem for Genomic Integration}
\usage{
.gint_subItem_HeatmapGenomic(data_table)
}
\description{
Heatmap Subitem for Genomic Integration
}
