% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_networkDEGs}
\alias{.gint_subItem_networkDEGs}
\title{Network Subitem for class comparison}
\usage{
.gint_subItem_networkDEGs(data_table)
}
\description{
Network Subitem for class comparison
}
