% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_networkTranscript}
\alias{.gint_subItem_networkTranscript}
\title{Network Subitem for Transcription Integration}
\usage{
.gint_subItem_networkTranscript(data_table)
}
\description{
Network Subitem for Transcription Integration
}
