% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_shiny.R
\name{.gint_subItem_tableCompleteInt}
\alias{.gint_subItem_tableCompleteInt}
\title{Table Subitem with all integrations}
\usage{
.gint_subItem_tableCompleteInt(data_table)
}
\description{
Table Subitem with all integrations
}
