% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.prepare_cnv_heatmap}
\alias{.prepare_cnv_heatmap}
\title{Prepare CNV Heatmap}
\usage{
.prepare_cnv_heatmap(
  data_table,
  df_heatmap,
  df_heatmap_t,
  significativityCriteria,
  pvalRange,
  fdrRange,
  numTopCNVonly,
  scale
)
}
\description{
Prepare CNV Heatmap
}
