% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.prepare_gen_heatmap}
\alias{.prepare_gen_heatmap}
\title{Prepare Genomic Heatmap}
\usage{
.prepare_gen_heatmap(
  data_table,
  df_heatmap,
  df_heatmap_t,
  significativityCriteria,
  pvalRange,
  fdrRange,
  numTopCNV,
  numTopMET,
  scale
)
}
\description{
Prepare Genomic Heatmap
}
