% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.prepare_met_heatmap}
\alias{.prepare_met_heatmap}
\title{Prepare Met Heatmap}
\usage{
.prepare_met_heatmap(
  data_table,
  df_heatmap,
  df_heatmap_t,
  significativityCriteria,
  pvalRange,
  fdrRange,
  numTopMETonly,
  scale
)
}
\description{
Prepare Met Heatmap
}
