% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.prepare_mirna_heatmap}
\alias{.prepare_mirna_heatmap}
\title{Prepare miRNA Heatmap}
\usage{
.prepare_mirna_heatmap(
  data_table,
  df_heatmap,
  df_heatmap_t,
  significativityCriteria,
  pvalRange,
  fdrRange,
  numTopMiCNV,
  scale
)
}
\description{
Prepare miRNA Heatmap
}
