% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.prepare_reactive_heatmap}
\alias{.prepare_reactive_heatmap}
\title{Prepare Reactive Heatmap}
\usage{
.prepare_reactive_heatmap(
  data_table,
  multiomics_integration,
  input,
  output,
  session,
  ns,
  deg = FALSE
)
}
\description{
Prepare Reactive Heatmap
}
