% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.prepare_reactive_histo}
\alias{.prepare_reactive_histo}
\title{Prepare Reactive Histogram}
\usage{
.prepare_reactive_histo(data_table, input, output, deg = FALSE, table = FALSE)
}
\description{
Prepare Reactive Histogram
}
