% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.prepare_reactive_histo_tf}
\alias{.prepare_reactive_histo_tf}
\title{Prepare Reactive Histogram for TF}
\usage{
.prepare_reactive_histo_tf(data_table, input, output, deg = FALSE)
}
\description{
Prepare Reactive Histogram for TF
}
