% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.prepare_reactive_ridge}
\alias{.prepare_reactive_ridge}
\title{Prepare Reactive Ridge Plot}
\usage{
.prepare_reactive_ridge(data_table, input, output, deg = FALSE, table = FALSE)
}
\description{
Prepare Reactive Ridge Plot
}
