% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_shiny.R
\name{.prepare_reactive_volcano}
\alias{.prepare_reactive_volcano}
\title{Prepare Reactive Volcano}
\usage{
.prepare_reactive_volcano(data_table, input, output, deg = FALSE)
}
\description{
Prepare Reactive Volcano
}
