% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_shiny.R
\name{.render_histo_tf_table}
\alias{.render_histo_tf_table}
\title{Render histogram for TFs/miRNAs as a table}
\usage{
.render_histo_tf_table(reactive_histo_tf_table)
}
\description{
Render histogram for TFs/miRNAs as a table
}
