% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_models.R
\name{.run_edgeR_integration}
\alias{.run_edgeR_integration}
\title{Running edgeR integration}
\usage{
.run_edgeR_integration(
  response_var,
  covariates,
  interactions = "auto",
  design_mat_allgene = NULL,
  offset_allgene = NULL,
  offset_singlegene = NULL,
  normalize = TRUE,
  norm_method = "TMM",
  steady_covariates = NULL,
  reference = NULL,
  BPPARAM = SerialParam()
)
}
\description{
Running edgeR integration
}
