% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{.run_lm_integration}
\alias{.run_lm_integration}
\title{Running linear model integration}
\usage{
.run_lm_integration(
  response_var,
  covariates,
  interactions = "auto",
  steady_covariates = NULL,
  reference = NULL,
  normalize = TRUE,
  norm_method = "TMM",
  BPPARAM = SerialParam()
)
}
\description{
Running linear model integration
}
