% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_shiny.R
\name{.server_histo}
\alias{.server_histo}
\title{Server Genomic integration chr distribution}
\usage{
.server_histo(input, output, session, data_table)
}
\description{
Server Genomic integration chr distribution
}
