% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_shiny.R
\name{.server_histo2}
\alias{.server_histo2}
\title{Server transcriptional and deg integration}
\usage{
.server_histo2(input, output, session, data_table, deg = FALSE)
}
\description{
Server transcriptional and deg integration
}
