% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_models.R
\name{.singlegene_edgeR_model}
\alias{.singlegene_edgeR_model}
\title{Single gene edgeR model}
\usage{
.singlegene_edgeR_model(
  response_var,
  covariates,
  offset_singlegene = NULL,
  fformula,
  fit_all,
  BPPARAM
)
}
\description{
Single gene edgeR model
}
