% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dot_plotly}
\alias{dot_plotly}
\title{plotting enrichment}
\usage{
dot_plotly(
  enrich_result,
  title = NULL,
  showCategory = 10,
  width = 800,
  height = 700
)
}
\arguments{
\item{enrich_result}{Enrichment analysis results.}

\item{title}{Title of the plot.}

\item{showCategory}{Number of categories to display.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\value{
A plotly object containing the dot plot.
}
\description{
plotting enrichment
}
\examples{
# Example usage:
# dot_plotly(enrichment_result, title = "Enrichment Analysis",
showCategory = 10)

}
