% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_models.R
\name{edger_coef_test}
\alias{edger_coef_test}
\title{Perform likelihood ratio test for all the coefficients
contained in the edgeR fitted model}
\usage{
edger_coef_test(fit_list, BPPARAM)
}
\arguments{
\item{fit}{A fitted edgeR model}

\item{threads}{Number of threads to use}
}
\value{
A data frame containing the results of the likelihood ratio test
for all the coefficients present in the model
}
\description{
Perform likelihood ratio test for all the coefficients
contained in the edgeR fitted model
}
