% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_model_res,list-method}
\alias{extract_model_res,list-method}
\title{Setting method for extracting results}
\usage{
\S4method{extract_model_res}{list}(
  model_results,
  outliers = TRUE,
  species = "hsa",
  filters = c("hgnc_symbol", "ensembl_gene_id", "entrezgene_id"),
  genes_info = NULL,
  ...
)
}
\description{
Setting method for extracting results
}
