% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/helper_functions.R
\name{extract_model_res}
\alias{extract_model_res}
\alias{extract_model_res,list-method}
\alias{extract_model_res,MultiClass-method}
\alias{extract_model_res,MultiOmics-method}
\title{Setting method for extracting results}
\usage{
extract_model_res(model_results, ...)

\S4method{extract_model_res}{list}(
  model_results,
  outliers = TRUE,
  species = "hsa",
  filters = c("hgnc_symbol", "ensembl_gene_id", "entrezgene_id"),
  genes_info = NULL,
  ...
)

\S4method{extract_model_res}{MultiClass}(
  model_results,
  outliers = TRUE,
  species = "hsa",
  filters = c("hgnc_symbol", "ensembl_gene_id", "entrezgene_id"),
  genes_info = NULL,
  ...
)

\S4method{extract_model_res}{MultiOmics}(
  model_results,
  outliers = TRUE,
  species = "hsa",
  filters = c("hgnc_symbol", "ensembl_gene_id", "entrezgene_id"),
  genes_info = NULL,
  ...
)
}
\arguments{
\item{model_results}{The model results object from which to extract results.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
A dataframe containing the results of all the integration models
provided
}
\description{
Setting method for extracting results
}
\examples{
# example code
data("ov_test_tcga_omics")
multiomics_integration <- run_multiomics(data = mmultiassay_ov)
data_table <- extract_model_res(multiomics_integration)
head(data_table)
}
