% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gINTomics-package.R
\docType{package}
\name{gINTomics-package}
\alias{gINTomics}
\alias{gINTomics-package}
\title{gINTomics: Multi-Omics data integration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

gINTomics is an R package for Multi-Omics data integration and visualization. gINTomics is designed to detect the association between the expression of a target and of its regulators, taking into account also their genomics modifications such as Copy Number Variations (CNV) and methylation. What is more, gINTomics allows integration results visualization via a Shiny-based interactive app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/angelovelle96/gINTomics}
  \item Report bugs at \url{https://github.com/angelovelle96/gINTomics/issues}
}

}
\author{
\strong{Maintainer}: Angelo Velle \email{angelo.velle@unipd.it} (\href{https://orcid.org/0000-0002-4010-6390}{ORCID})

Authors:
\itemize{
  \item Francesco Patane' \email{francesco.patane@unipd.it} (\href{https://orcid.org/0009-0001-8619-447X}{ORCID})
  \item Chiara Romualdi \email{chiara.romualdi@unipd.it} (\href{https://orcid.org/0000-0003-4792-9047}{ORCID})
}

}
\keyword{internal}
