% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirna_hsa}
\alias{mirna_hsa}
\title{miRNA IDs.
Dataset containing lastly definition of miRNAs (Names, Accessions,
Sequences, Families and others) from different miRBase versions
(From miRBase version 6 to version 22).}
\usage{
data(miRBase_conversion)
"mirna_hsa"
}
\value{
An object of class \linkS4class{data.frame}.
}
\description{
miRNA IDs.
Dataset containing lastly definition of miRNAs (Names, Accessions,
Sequences, Families and others) from different miRBase versions
(From miRBase version 6 to version 22).
}
\examples{
# example code
data(miRBase_conversion)
head(mirna_hsa)
}
