% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmultiassay_ov}
\alias{mmultiassay_ov}
\title{Example data for a standard workflow.
This is an example dataset containing a MultiAssayExperiment of 20 ovarian
cancer (OVC) patients extracted from the Cancer Genome Atlas (TCGA) database.
The object contains all the available input data types: Gene expression data,
miRNA expression data, gene methylation data, gene Copy Number Variations
and miRNA Copy Number Variations.}
\format{
MultiAssayExperiment
}
\usage{
data(mmultiassay_ov)
"mmultiomics_ov"
}
\value{
An object of class \linkS4class{MultiAssayExperiment}.
}
\description{
Example data for a standard workflow.
This is an example dataset containing a MultiAssayExperiment of 20 ovarian
cancer (OVC) patients extracted from the Cancer Genome Atlas (TCGA) database.
The object contains all the available input data types: Gene expression data,
miRNA expression data, gene methylation data, gene Copy Number Variations
and miRNA Copy Number Variations.
}
\examples{
# example code
data(mmultiassay_ov)
mmultiassay_ov
}
