% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MultiClass-class}
\alias{MultiClass-class}
\title{MultiClass Class}
\value{
MultiOmics Class
}
\description{
S4 class containing the output of a single integration integration, for
which classes has been provided. It's a list in which each element
represents the result of the integration for a given class. The length will
be equal to the number of classes defined.
}
