% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MultiOmics-class}
\alias{MultiOmics-class}
\title{MultiOmics Class}
\value{
MultiOmics Class
}
\description{
S4 class containing the output of a multiomics integration. It's a list in
which each element represents the result of an integration. If all the
available omics are provided, it will be a list of integrations:
\strong{gene_genomic_res}, \strong{mirna_cnv_res}, \strong{tf_res}, \strong{tf_mirna_res} and
\strong{mirna_target_res}
}
