% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_chr_distribution}
\alias{plot_chr_distribution}
\title{plotting chr distribution}
\usage{
plot_chr_distribution(
  data_table,
  class = NULL,
  omics = NULL,
  cnv_met = NULL,
  pval = 0.05
)
}
\description{
plotting chr distribution
}
