% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_network}
\alias{plot_network}
\title{Plotting network}
\usage{
plot_network(data_table, num_interactions = 300, pval = 0.05)
}
\arguments{
\item{data_table}{The data table containing network information.}

\item{num_interactions}{The number of interactions to display in the network
(default: 300).}

\item{pval}{The p-value threshold for selecting interactions (default: 0.05).}
}
\value{
A network plot.
}
\description{
Plotting network
}
\examples{
# Example usage:
library(MultiAssayExperiment)
data("mmultiassay_ov")
tmp <- lapply(experiments(mmultiassay_ov), function(x) x[1:20,])
mmultiassay_ov <- MultiAssayExperiment(experiments = tmp)
# multiomics_integration <- run_multiomics(data = mmultiassay_ov)
# data_table <- extract_model_res(multiomics_integration)
# plot_network(data_table)
}
