% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tf_distribution}
\alias{plot_tf_distribution}
\title{plotting TF distribution}
\usage{
plot_tf_distribution(data_table, pval = 0.05)
}
\arguments{
\item{data_table}{The data table containing TF information.}

\item{pval}{Optional. The p-value threshold for significance in the
distribution plot. Default is 0.05.}
}
\value{
A TF distribution plot.
}
\description{
plotting TF distribution
}
\examples{
# Example usage:
library(MultiAssayExperiment)
data("mmultiassay_ov")
tmp <- lapply(experiments(mmultiassay_ov), function(x) x[1:20,])
mmultiassay_ov <- MultiAssayExperiment(experiments = tmp)
# multiomics_integration <- run_multiomics(data = mmultiassay_ov)
# data_table <- extract_model_res(multiomics_integration)
# plot_tf_distribution(data_table, pval=0.5)
}
