% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_venn}
\alias{plot_venn}
\title{plotting venn}
\usage{
plot_venn(data_table, class = NULL)
}
\arguments{
\item{data_table}{The data table containing information for the Venn diagram.}

\item{class}{Optional. The class of interactions to include in the Venn
diagram.}
}
\value{
A Venn diagram plot.
}
\description{
plotting venn
}
\examples{
# Example usage:
library(MultiAssayExperiment)
data("ov_test_tcga_omics")
tmp <- lapply(mmultiassay_ov@ExperimentList, function(x) x[1:20,])
mmultiassay_ov <- MultiAssayExperiment(experiments = tmp)
multiomics_integration <- run_multiomics(data = mmultiassay_ov)
data_table <- extract_model_res(multiomics_integration)
plot_venn(data_table)
}
