% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ridgeline_plot}
\alias{ridgeline_plot}
\title{Ridgeline plot}
\usage{
ridgeline_plot(data, outliers = F)
}
\arguments{
\item{data}{Output of one of the integration functions or output of
\strong{run_multiomics}}

\item{outliers}{logical. Should outliers be showed in the plot. Default is
set to FALSE}
}
\description{
The ridgeline plot allow to compare the distribution of significant and
non significant coefficients.
}
