% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{run_cnv_integration}
\alias{run_cnv_integration}
\title{Integration of expression and Copy Number Variations}
\usage{
run_cnv_integration(
  expression,
  cnv_data,
  sequencing_data = T,
  normalize = T,
  norm_method = "TMM",
  class = NULL,
  run_deg = TRUE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{expression}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models.}

\item{cnv_data}{Matrix or data.frame containing the Copy Number variation
status for the models. Rows represent samples, while columns represent
the different covariates. If \strong{interactions} are not provided, they will be
automatically generated and for each gene contained in \strong{expression}
the model will look for the same gene in \strong{cnv_data}}

\item{sequencing_data}{logical. Are expression data obtained from RNA
sequencing ? Default is set to TRUE}

\item{normalize}{logical.Should expression data be
normalized ? Default is set to TRUE}

\item{norm_method}{Normalization method to be used for
expression data. One of "TMM" (default), "TMMwsp", "RLE", "upperquartile",
"none".}
}
\description{
This function will perform an integration of expression data and Copy Number
Variations data
}
