% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edgeR_integration.R
\name{run_edgeR_integration}
\alias{run_edgeR_integration}
\title{Integration of omics data running single gene negative binomial edgeR models}
\usage{
run_edgeR_integration(
  response_var,
  covariates = NULL,
  interactions = "auto",
  design_mat_allgene = NULL,
  design_mat_singlegene = NULL,
  offset_allgene = NULL,
  offset_singlegene = NULL,
  norm_method = "TMM",
  steady_covariates = NULL,
  threads = 1,
  reference = NULL
)
}
\arguments{
\item{response_var}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models. The number of rows should
be identical to that of \strong{covariates} or of design matrices provided in
\strong{design_mat_singlegene} and the order of the samples should be the same.
In case you need to run a single model you can provide the response
variable as an atomic vector, anyway the response variables should always
be numeric values.}

\item{covariates}{Matrix or data.frame containing the covariates of the
models. Rows represent samples, while columns represent the different
covariates. This argument is ignored if \strong{design_mat_singlegene}
is provided.}

\item{interactions}{A list of characters containing the covariates to use
for each single gene model. Each element of the list should be a character
vector corresponding to colnames in \strong{covariates}, for each genes the
specified columns will be used to create the design matrix. The list should
contain the covariates for each of the genes of \strong{response_var} and they
should be in the same order of \strong{response_var} rownames.
This argument is ignored if \strong{design_mat_singlegene} is provided.}

\item{design_mat_allgene}{A design matrix for the all gene edgeR model}

\item{design_mat_singlegene}{A list of design matrices for
each edgeR single gene model, the number of design matrices should be equal
to the number of rows of \strong{response_var}.}

\item{offset_allgene}{A matrix containing the offsets of the all gene edgeR
model}

\item{offset_singlegene}{A list of vector containing the offsets for each
single gene edgeR model.
The number of vectors should be equal to the number of rows
of \strong{response_var}. As default, offsets are taken from the all gene fitting.}

\item{norm_method}{Normalization method to use for the all gene edgeR
model (Default is 'TMM'). The normalization factors will be passed to single
gene models for normalization.}

\item{steady_covariates}{Character vector containing column names of
\strong{covariates} corresponding to covariates that should be included in all
single gene models (default=NULL).}

\item{threads}{Number of threads to use for parallelization (Default is 1)}
}
\value{
A list containing the results of all the edger single gene models,
the pvalues and the coefficients for each gene
}
\description{
This function is useful for users which need to run single gene
integration models, in particular if for each gene the covariates change.
For example if you need to integrate gene expression with copy number
variations, for each gene you need to integrate its expression levels with
CNV status, updating the covariate at each step. In this case you can set
\strong{interactions} to auto (default), providing the expression of all genes in
\strong{response_var} and the relative CNV values (numeric) in \strong{covariates}.
The function will automatically use CNV values of each gene as covariates
for the single gene models.
The user can specify which covariates should be used for each gene in the
\strong{interaction} argument.
The user can manually provide a list of the design matrices for single gene
models in \strong{design_mat_singlegene}, in this case the n gene contained in
\strong{response_var} is fitted in the  model  using as design matrix the n
element of the\strong{design_mat_singlegene} list.
Since the function runs a model for each of the genes contained in
\strong{response_var}, in order to compute the common and tagwise dispersion it
starts running an all gene model without covariates, then the dispersion
values and the offsets are passed to the single gene model.
}
