% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{run_genomic_enrich}
\alias{run_genomic_enrich}
\title{Running genomic enrichment analysis}
\usage{
run_genomic_enrich(
  model_results,
  species = "hsa",
  pvalueCutoff = 0.1,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.1,
  ont = "all",
  BPPARAM = BiocParallel::SerialParam(),
  extracted_data = NULL,
  ...
)
}
\arguments{
\item{qvalueCutoff}{description}

\item{BPPARAM}{description}

\item{extracted_data}{description}

\item{...}{}
}
\description{
Running genomic enrichment analysis
}
