% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{run_met_integration}
\alias{run_met_integration}
\title{Integration of expression and methylation}
\usage{
run_met_integration(
  expression,
  methylation,
  sequencing_data = T,
  normalize = T,
  norm_method = "TMM",
  class = NULL,
  run_deg = T,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{expression}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models.}

\item{methylation}{Matrix or data.frame containing the methylation
values for the models. Rows represent samples, while columns represent
the different covariates. If \strong{interactions} are not provided, they will be
automatically generated and for each gene contained in \strong{expression}
the model will look for the same gene in \strong{methylation}}

\item{sequencing_data}{logical. Are expression data obtained from RNA
sequencing ? Default is set to TRUE}

\item{normalize}{logical.Should expression data be
normalized ? Default is set to TRUE}

\item{norm_method}{Normalization method to be used for
expression data. One of "TMM" (default), "TMMwsp", "RLE", "upperquartile",
"none".}
}
\description{
This function will perform an integration of expression data and methylation
data
}
