% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_singlegene_edgeR_integration.R
\name{run_singlegene_edgeR_integration}
\alias{run_singlegene_edgeR_integration}
\title{Integration of omics data running single gene edgeR models}
\usage{
run_singlegene_edgeR_integration(
  response_var,
  covariates = NULL,
  design_mat_allgene = NULL,
  design_mat_singlegene = NULL,
  offset_allgene = NULL,
  offset_singlegene = NULL,
  norm_method = "TMM",
  threads = 1
)
}
\arguments{
\item{response_var}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models. The number of rows should
be identical to that of \strong{covariates} or of design matrices provided in
\strong{design_mat_singlegene} and the order of the samples should be the same.
In case you need to run a single model you can provide the response
variable as an atomic vector, anyway the response variables should always
be numeric values.}

\item{covariates}{Matrix or data.frame containing the covariates of the
models. Rows represent samples, while columns represent the different
the covariates. All the covariates contained in columns will be used for
each model, so use this argument only if all the models have the same
covariates. In case you need a single covariate you can provide the
covariate as an atomic vector. This argument is ignored if
\strong{design_mat_singlegene} is provided.}

\item{design_mat_allgene}{A design matrix for the all gene edgeR model}

\item{design_mat_singlegene}{A design matrix (if the model to run is one) or
a list of design matrices for each edgeR single gene model. The number of
design matrices should be equal to the number of rows of \strong{response_var}.}

\item{offset_allgene}{A matrix containing the offsets of the all gene edgeR
model}

\item{offset_singlegene}{A vector (if the model to run is one) or
a list of vector containing the offsets for each single gene edgeR model.
The number of vectors should be equal to the number of rows
of \strong{response_var}.}

\item{norm_method}{Normalization method to use for the all gene edgeR
model (Default is 'TMM'). The normalization factors will be passed to single
gene models for normalization. Ignored if the offsets are provided.}

\item{threads}{Number of threads to use for parallelization (Default is 1)}
}
\value{
A list containing the results of all the edger single gene models,
the pvalues and the coefficients for each gene
}
\description{
This function is useful for users which need to run single gene
integration models, in particular if for each gene the covariates change.
For example if you need to integrate gene expression with copy number
variations, for each gene you need to integrate its expression levels with
CNV status, updating the covariate at each step. In this case you can
provide the expression of all genes in \strong{response_var} and the relative
CNV values (numeric) as a list of design matrices in
\strong{design_mat_singlegene}.
The n gene contained in \strong{response_var} is fitted in the negative binomial
model of the edgeR package using as design matrix the n element of the
\strong{design_mat_singlegene} list. Alternatively, if the covariates are the
same for each gene, the user can provide them as a dataframe in
\strong{covariates}, in this case all the columns will be included in the design
matrix. Since the function runs a model for each of the genes contained in
\strong{response_var}, in order to compute the common and tagwise dispersion it
starts running an all gene model without covariates, then the dispersion
values are passed to the single gene model.
}
