% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlegene_edgeR_model.R
\name{singlegene_edgeR_model}
\alias{singlegene_edgeR_model}
\title{Single gene omics integration}
\usage{
singlegene_edgeR_model(
  response_var,
  interactions = NULL,
  covariates = NULL,
  offset_singlegene = NULL,
  design_mat = NULL,
  y_all,
  threads = 1
)
}
\arguments{
\item{response_var}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models. The number of rows should
be identical to that of \strong{covariates} or of design matrices provided in
\strong{design_mat_singlegene} and the order of the samples should be the same.
In case you need to run a single model you can provide the response
variable as an atomic vector, anyway the response variables should always
be numeric values.}

\item{interactions}{A list of characters containing the covariates to use
for each single gene model. Each element of the list should be a character
vector corresponding to colnames in \strong{covariates}, for each genes the
specified columns will be used to create the design matrix.
This argument is ignored if \strong{design_mat_singlegene} is provided.}

\item{covariates}{Matrix or data.frame containing the covariates of the
models. Rows represent samples, while columns represent the different
covariates. This argument is ignored if \strong{design_mat_singlegene}
is provided.}

\item{offset_singlegene}{A vector (if the model to run is one or if the
offsets don't change across models) or a list of vectors containing the
offsets for each single gene edgeR model.
If provided as list, the number of vectors should be equal to the number
of rows of \strong{response_var}.}

\item{design_mat}{A design matrix (if the model to run is one or if the
matrix doesn't change across models) or a list of design matrices for each
edgeR single gene model.
If provided as list, the number of design matrices should be equal to the
number of rows of \strong{response_var}.}

\item{y_all}{An all gene edgeR model needed to extract the common and
tagwise dispersion}

\item{threads}{Number of threads to use for parallelization (Default is 1)}
}
\description{
Compute single gene edgeR's models for each of the genes
provided in \strong{response_var}
}
