% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_chr_distribution}
\alias{plot_chr_distribution}
\title{plotting chr distribution}
\usage{
plot_chr_distribution(
  data_table,
  class = NULL,
  omics = NULL,
  cnv_met = NULL,
  pval = 0.05
)
}
\arguments{
\item{data_table}{The data table containing information for plotting
chromosome distribution.}

\item{class}{Optional. The class of interactions to include in the plot.}

\item{omics}{Optional. The type of omics data for the plot.}

\item{cnv_met}{Optional. The type of copy number variation or methylation
data.}

\item{pval}{Optional. The p-value threshold for significance. Default is
0.05.}
}
\value{
A histogram plot showing chromosome distribution.
}
\description{
plotting chr distribution
}
\examples{
# Example usage:
data("ov_test_tcga_omics")
multiomics_integration <- run_multiomics(data = mmultiassay_ov)
data_table <- extract_model_res(multiomics_integration)
plot_chr_distribution(data_table)
}
