% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{run_genomic_integration}
\alias{run_genomic_integration}
\title{Integration of expression, Copy Number Variations and methylation data}
\usage{
run_genomic_integration(
  expression,
  cnv_data,
  methylation,
  sequencing_data = T,
  normalize = T,
  norm_method = "TMM",
  interactions = NULL,
  class = NULL,
  scale = T,
  run_deg = T,
  ...
)
}
\arguments{
\item{expression}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models.}

\item{cnv_data}{Matrix or data.frame containing the Copy Number variation
status for the models. Rows represent samples, while columns represent
the different covariates. If \strong{interactions} are not provided, they will be
automatically generated and for each gene contained in \strong{expression}
the model will look for the same gene in \strong{cnv_data}}

\item{methylation}{Matrix or data.frame containing the methylation
values for the models. Rows represent samples, while columns represent
the different covariates. If \strong{interactions} are not provided, they will be
automatically generated and for each gene contained in \strong{expression}
the model will look for the same gene in \strong{methylation}
#' @param sequencing_data logical. Are expression data obtained from RNA
sequencing ? Default is set to TRUE}

\item{normalize}{logical.Should expression data be
normalized ? Default is set to TRUE}

\item{norm_method}{Normalization method to be used for
expression data. One of "TMM" (default), "TMMwsp", "RLE", "upperquartile",
"none".}

\item{interactions}{A list of character vectors containing the interactions
between response variable and covariates. The names of the list should
match the response variables while the character contained in each element
of the list should match the covariates. If NULL (default), the interactions
will be automatically defined according to response variable's colnames.}
}
\description{
This function will perform an integration of expression data and Copy Number
Variations data
}
