% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{run_tf_enrich}
\alias{run_tf_enrich}
\title{Running TF enrichment analysis}
\usage{
run_tf_enrich(
  model_results,
  species = "hsa",
  pvalueCutoff = 0.1,
  qvalueCutoff = 0.1,
  pAdjustMethod = "BH",
  ont = "all",
  BPPARAM = BiocParallel::SerialParam(),
  extracted_data = NULL,
  ...
)
}
\arguments{
\item{model_results}{Model integration results, typically a list containing
TF data.}

\item{species}{Species to select for the enrichment analysis. Default is
'hsa' (Homo sapiens).}

\item{pvalueCutoff}{P-value cutoff for significant enrichment. Default is
0.1.}

\item{qvalueCutoff}{Q-value cutoff for significant enrichment. Default is
0.1.}

\item{pAdjustMethod}{Method for adjusting p-values. Default is 'BH'
(Benjamini & Hochberg).}

\item{ont}{Ontology to use for the enrichment analysis. Default is 'all'.}

\item{BPPARAM}{A BiocParallelParam object specifying parallelization
options. Default is BiocParallel::SerialParam().}

\item{extracted_data}{Pre-extracted data for enrichment analysis. If NULL,
function will extract relevant data from model_results.}

\item{...}{Additional arguments to be passed to the internal enrichment
function.}
}
\value{
A list containing TF enrichment results.
}
\description{
Running TF enrichment analysis
}
\examples{
# Example usage:
library(MultiAssayExperiment)
data(ov_test_tcga_omics)
tmp <- lapply(mmultiassay_ov@ExperimentList, function(x) x[1:200,])
mmultiassay_ov <- MultiAssayExperiment(experiments = tmp)
#multiomics_integration <- run_multiomics(mmultiassay_ov)
#run_tf_enrich(multiomics_integration, qvalueCutoff = 1, pvalueCutoff = 0.05,
#pAdjustMethod = 'none')
}
