% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{run_tf_enrich}
\alias{run_tf_enrich}
\title{Title}
\usage{
run_tf_enrich(
  model_results,
  species = "hsa",
  pvalueCutoff = 0.1,
  pAdjustMethod = "BH",
  ont = "all",
  BPPARAM = BiocParallel::SerialParam(),
  extracted_data = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
