% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{run_tf_integration}
\alias{run_tf_integration}
\title{Integration of expression and Transcription Factors / Generic Regulators}
\usage{
run_tf_integration(
  expression,
  tf_expression = expression,
  interactions = NULL,
  type = "none",
  sequencing_data = T,
  species = "hsa",
  normalize = T,
  norm_method = "TMM",
  normalize_cov = T,
  norm_method_cov = "TMM",
  ...
)
}
\arguments{
\item{expression}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models.}

\item{tf_expression}{Matrix or data.frame containing the expression
values for the models. Rows represent samples, while columns represent
the different covariates. If not provided, it will be set equal to
\strong{expression}.}

\item{interactions}{A list of character vectors containing the interactions
between response variable and covariates. The names of the list should
match the response variables while the character contained in each element
of the list should match the covariates. If NULL (default), the interactions
will be automatically downloaded according to the \strong{type} argument.}

\item{type}{A character defining the type of regulation under analysis.
Should be one of "tf_miRNA", "tf", "miRNA_target".}

\item{sequencing_data}{logical. Are expression data obtained from RNA
sequencing ? Default is set to TRUE}

\item{species}{species information for interactions download. Fully
supported species are "hsa" (default) and "mmu".}

\item{normalize}{logical.Should expression data be
normalized ? Default is set to TRUE}

\item{norm_method}{Normalization method to be used for
expression data. One of "TMM" (default), "TMMwsp", "RLE", "upperquartile",
"none".}

\item{normalize_cov}{Same as \strong{normalize} but for covariates.}

\item{norm_method_cov}{Same as \strong{norm_method} but for covariates.}
}
\description{
This function will perform an integration of gene/miRNA expression data
and Transcription Factors expression. Moreover, every type of regulator can
be provided to the function as covariate through the \strong{tf_expression}
argument. Interactions for TF-target, miRNA-target and TF-miRNA integration
will be automatically downloaded by the function as defined by the \strong{type}
argument. Other types of interactions should be provided through the
\strong{interactions} argument.
}
