% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlegene_edgeR_model.R
\name{singlegene_edgeR_model}
\alias{singlegene_edgeR_model}
\title{Single gene omics integration}
\usage{
singlegene_edgeR_model(
  response_var,
  covariates = NULL,
  design_mat = NULL,
  offset_singlegene = NULL,
  y_all,
  threads = 1
)
}
\arguments{
\item{response_var}{Matrix or data.frame containing the expression values
for each model. Rows represent samples, while each column represents
the different response variables of the models. The number of rows should
be identical to that of \strong{covariates} or of design matrices provided in
\strong{design_mat_singlegene} and the order of the samples should be the same.
In case you need to run a single model you can provide the response
variable as an atomic vector, anyway the response variables should always
be numeric values.}

\item{covariates}{Matrix or data.frame containing the covariates of the
models. Rows represent samples, while columns represent the different
the covariates. All the covariates contained in columns will be used for
each model, so use this argument only if all the models have the same
covariates. In case you need a single covariate you can provide the
covariate as an atomic vector. This argument is ignored if
\strong{design_mat_singlegene} is provided.}

\item{design_mat}{A design matrix (if the model to run is one) or
a list of design matrices for each edgeR single gene model. The number of
design matrices should be equal to the number of rows of \strong{response_var}.}

\item{offset_singlegene}{A vector (if the model to run is one) or
a list of vector containing the offsets for each single gene edgeR model.
The number of vectors should be equal to the number of rows
of \strong{response_var}.}

\item{y_all}{An all gene edgeR model needed to extract the common and
tagwise dispersion}

\item{threads}{Number of threads to use for parallelization (Default is 1)}
}
\description{
Compute single gene edgeR's models for each of the genes
provided in \strong{response_var}
}
